load in fp16 (less memory, faster and similar results)

+++ nlu/models.py
@@ -197,7 +197,7 @@ class UniversalModel():
 
     def __init__(self, classifier_path: str, extractor_path: str, device: str = 'cuda') -> None:
         peft_config = PeftConfig.from_pretrained(classifier_path)
-        self.model = RobertaForNLU.from_pretrained(peft_config.base_model_name_or_path, num_labels=3).eval()
+        self.model = RobertaForNLU.from_pretrained(peft_config.base_model_name_or_path, num_labels=3, torch_dtype=torch.float16).eval()
         self.model = PeftForNLU.from_pretrained(model=self.model, model_id=classifier_path, adapter_name='classification').to(device)
         self.model.load_adapter(model_id=extractor_path, adapter_name='extraction')
         self.tokenizer = RobertaTokenizerFast.from_pretrained(classifier_path)
